Allow returning results from the earlier two SDL functions
Any particular invocation of Screen::ResizeScreen() will end up calling
three SDL functions, but there can only be one error code returned. My
previous patch prioritized the last SDL function at all times, even if
the earlier functions failed.

So if the third fails, return the result of the third. Otherwise, if the
second fails, return the result of the second. Else just return the
result of the first.

This is in sync with the fact that SDL_GetError() will prioritize the
last failed function called, so the error code given is the one
corresponding with SDL_GetError() in the end.