Add mx and my attributes to entclass
Entities now have new attributes mx and my. These are the magnitude for
their x and y velocity, respectively, and are floats between 0 and 1
(although I clamp it down to -1 and 1 to allow for backwards movement if
necessary). They're essentially a percentage applied to the velocity
when newxp/newyp is computed, so you can never go higher than max
velocity. However, only mx will end up being used in the end.

This sidesteps directly setting the player's velocity or acceleration,
because due to the friction system it will set any velocity below 1.1 to
0, and it's less risk of changing the underlying physics and thus
breaking compatibility with levels. I'm pretty sure compatibility is
still kept either way.