Add analogue support for controller joystick
Analogue input support means you can now, with a joystick, have a choice
in between "full speed ahead" and "no movement", unlike using a
keyboard.

I've added a variable xMagnitude to KeyPoll, which should be a value
between -1 and 1. It calculates exactly how far left or right the
joystick is tilted, accounting for the configured controller sensitivity
so as to make sure all subtle values are still available even if you
have low sensitivity.